for v0.0.22

+++ EffectInfoBackend/Main.cs
@@ -546,7 +546,7 @@ namespace EffectInfo
                 result += ToInfoAdd("皱纹", value, 2);
             }
             {
-                var value = CallPrivateMethod<double>(avatar, "GetClothCharm", new object[] { (byte)clothingDisplayId });
+                var value = CallPrivateMethod<double>(avatar, "GetClothCharm", new object[] { clothingDisplayId });
                 result += ToInfoAdd("衣服", value, 2);
             }
             _headAsset = null;
@@ -855,7 +855,7 @@ namespace EffectInfo
                 {
                     short clothingDisplayId = character.GetClothingDisplayId();
                     //正式版这里clothingDisplayId转成byte明显是bug，而测试版修复了
-                    check_value = (int)character.GetAvatar().GetCharm(physiologicalAge, (byte)clothingDisplayId);
+                    check_value = (int)character.GetAvatar().GetCharm(physiologicalAge, clothingDisplayId);
                     if (check_value != 0 || ShowUseless)
                     {
                         result += ToInfoAdd("外观", check_value, 1);
